Extract wireModMacroCallbacks template + icon/SvgButton polish
- Extract duplicate mod/macro callback wiring into DeviceLinkCallbacks.hpp
  template, eliminating ~490 lines of duplication between constructor
  param slot setup and setupCustomUILinking()
- Add float iconPadding to SvgButton for sub-pixel icon sizing control
- Set originalColor on multiout/open_in_new buttons for proper recoloring
- Reduce icon transparency to 50% alpha for better visual blending

+++ magda/daw/ui/components/common/SvgButton.hpp
@@ -52,7 +52,7 @@ class SvgButton : public juce::Button {
     }
 
     // Border settings
-    void setIconPadding(int padding) {
+    void setIconPadding(float padding) {
         iconPadding = padding;
     }
 
@@ -100,7 +100,7 @@ class SvgButton : public juce::Button {
     float borderThickness = 1.0f;
     float cornerRadius = 2.0f;
     bool hasBorder = false;
-    int iconPadding = 4;
+    float iconPadding = 4.0f;
 
     bool active = false;
